extended Person view and improved person details
