+++ components/PersonView/PersonView.tsx
@@ -1,11 +1,22 @@
-import PropTypes from 'prop-types';
-
 import ErrorMessage from 'components/ErrorMessage/ErrorMessage';
 import Spinner from 'components/Spinner/Spinner';
 import PersonDetails from './PersonDetails';
 import { useResident } from 'utils/api/residents';
+import { Resident } from 'types';
 
-const PersonView = ({ personId, expandView, children }) => {
+interface Props {
+  personId: number;
+  children?: React.ReactChild | ((arg0: Resident) => React.ReactChild);
+  expandView?: boolean;
+  showPersonDetails?: boolean;
+}
+
+const PersonView = ({
+  personId,
+  expandView,
+  children,
+  showPersonDetails = true,
+}: Props): React.ReactElement => {
   const { data: person, error } = useResident(personId);
   if (error) {
     return <ErrorMessage />;
@@ -20,16 +31,12 @@ const PersonView = ({ personId, expandView, children }) => {
           {person.firstName} {person.lastName}
         </h1>
       )}
-      <PersonDetails {...person} expandView={expandView} />
+      {showPersonDetails && (
+        <PersonDetails person={person} expandView={expandView} />
+      )}
       {typeof children === 'function' ? children(person) : children}
     </>
   );
-};
-
-PersonView.propTypes = {
-  expandView: PropTypes.bool,
-  personId: PropTypes.string.isRequired,
-  children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
 };
 
 export default PersonView;
