+++ pages/people/[id]/index.tsx
@@ -3,10 +3,9 @@ import { useRouter } from 'next/router';
 import Seo from 'components/Layout/Seo/Seo';
 import BackButton from 'components/Layout/BackButton/BackButton';
 import PersonView from 'components/PersonView/PersonView';
+import PersonDetails from 'components/PersonView/PersonDetails';
 import Cases from 'components/Cases/Cases';
 import AllocatedWorkers from 'components/AllocatedWorkers/AllocatedWorkers';
-import { Resident } from 'types';
-import React from 'react';
 import WarningNotes from 'components/WarningNote/WarningNotes';
 
 const PersonPage = (): React.ReactElement => {
@@ -20,6 +19,7 @@ const PersonPage = (): React.ReactElement => {
         {(person) => (
           <>
             <WarningNotes id={personId} />
+            <PersonDetails person={person} />
             <AllocatedWorkers id={personId} />
             <Cases id={personId} person={person} />
           </>
